refactor: move process_text into parsing
