+++ csv_detective/detect_fields/FR/other/insee_ape700/__init__.py
@@ -1,5 +1,5 @@
 from os.path import dirname, join
-from csv_detective.process_text import _process_text
+from csv_detective.parsing.text import _process_text
 
 PROPORTION = 1
 f = open(join(dirname(__file__), 'insee_ape700.txt'), 'r')
