+++ csv_detective/detect_fields/FR/other/sexe/__init__.py
@@ -1,4 +1,4 @@
-from csv_detective.process_text import _process_text
+from csv_detective.parsing.text import _process_text
 
 PROPORTION = 1
